Fix bug in #single_move_set & #single_capture_set
Previously, these methods incorrectly called the methods #create_moves
and #create_captures,resulting in the addition of invalid moves and
captures to their respective lists.

This commit corrects the issue by updating the calls to refer to the
appropriate methods, #create_single_moves and #create_single_captures,
ensuring the accuracy of moves and captures.